(svn r16188) -Fix [FS#2874] (r16124): printing the first 4 bytes worth of a 8 byte integer causes some unwanted side effects when the system is not little-endian.

+++ src/strings.cpp
@@ -203,7 +203,7 @@ static char *FormatNumber(char *buff, int64 number, const char *last, const char
 			num = num % divisor;
 		}
 		if (tot |= quot || i == 19) {
-			buff += seprintf(buff, last, "%i", quot);
+			buff += seprintf(buff, last, "%i", (int)quot);
 			if ((i % 3) == 1 && i != 19) buff = strecpy(buff, separator, last);
 		}
